+++ src/app/components/events/NotEvent.jsx
